+++ R/themes.R
@@ -6,37 +6,29 @@
 #'
 #' Complete [ggplot2](http://ggplot2.tidyverse.org) themes.
 #'
-#' Both themes are based off of [ggplot2::theme_linedraw()], but with
-#' modifications and extra user-definable parameters.
-#'
-#' @section Light:
-#'
-#' High contrast black and white theme optimized for print. Recommended for
-#' scientific manuscripts and website tutorials.
-#'
-#' @section Dark:
-#'
-#' Blackout theme that sets the plot background as black, with white text.
-#' Inspired by `Seurat::DarkTheme()`, with some color modifications. Useful
-#' for visualizing many points with a high dynamic color range, such as t-SNE
-#' expression plots.
+#' Themes are based off of [ggplot2::theme_linedraw()], but with modifications
+#' and extra user-definable parameters.
 #'
 #' @name themes
-#' @note Updated 2019-07-29.
+#' @note Updated 2019-09-13.
 #'
-#' @param base_size `numeric(1)`. Base font size.
-#' @param base_family `character(1)`. Base font family.
-#' @param face `character(1)`. Font face (`"bold"`, `"plain"`).
-#' @param aspect_ratio `numeric(1)`. Aspect ratio, specifying the plot
-#'   proportions. Use `1` for a perfectly square plot (including the axis
-#'   labels).
-#' @param legend_position `character(1)`. Legend key position. We're being a
-#'   little more restrictive here, only allowing `"bottom"`, `"right"`, or
-#'   `"none"`. Including the legend at the top or the left side of the plot
-#'   rarely makes sense and is discouraged.
-#' @param grid `logical(1)`. Label the major panel grids with a gray accent.
-#' @param minimal `logical(1)`. Remove all axis lines, axis ticks, and
-#'   panel borders.
+#' @param base_size `numeric(1)`.
+#'   Base font size.
+#' @param base_family `character(1)`.
+#'   Base font family.
+#' @param face `character(1)`.
+#'   Font face (`"bold"`, `"plain"`).
+#' @param aspect_ratio `numeric(1)`.
+#'   Aspect ratio, specifying the plot proportions. Use `1` for a perfectly
+#'   square plot (including the axis labels).
+#' @param legend_position `character(1)`.
+#'   Legend key position. We're being a little more restrictive here, only
+#'   allowing `"bottom"`, `"right"`, or `"none"`. Including the legend at the
+#'   top or the left side of the plot rarely makes sense and is discouraged.
+#' @param grid `logical(1)`.
+#'   Label the major panel grids.
+#' @param minimal `logical(1)`.
+#'   Remove all axis lines, axis ticks, and panel borders.
 #'
 #' @return `theme`.
 #'
@@ -55,16 +47,21 @@
 #' ) +
 #'     geom_point()
 #'
-#' ## Paperwhite theme.
+#' ## Light theme.
 #' p + acid_theme_light(legend_position = "none")
 #'
-#' ## Midnight theme.
+#' ## Dark theme.
 #' p + acid_theme_dark(legend_position = "none")
+#'
+#' ## Dracula theme.
+#' p + acid_theme_dracula(legend_position = "none")
 NULL
 
 
 
-#' @rdname themes
+#' @describeIn themes Light theme that has a white background and black text.\cr
+#'   Optimized for print and recommended for scientific manuscripts.\cr
+#'   See `lightPalette` for color values.
 #' @export
 acid_theme_light <-  # nolint
     function(
@@ -76,11 +73,13 @@ acid_theme_light <-  # nolint
         grid = FALSE,
         minimal = FALSE
     ) {
+        palette <- lightPalette
+        ## Set dark mode global variable that we can access inside functions.
+        options(acid.dark = FALSE)
         assert(
             isNumber(base_size),
-            ## Don't use `isString()` check on `base_family`, since empty string
-            ## is allowed by ggplot2.
-            is.character(base_family) && length(base_family) == 1L
+            is.character(base_family),
+            isScalar(base_family)
         )
         face <- match.arg(face)
         assert(isNumber(aspect_ratio, nullOK = TRUE))
@@ -89,15 +88,17 @@ acid_theme_light <-  # nolint
             isFlag(grid),
             isFlag(minimal)
         )
-        gray <- "gray95"
+        bg <- palette[["background"]]
+        fg <- palette[["foreground"]]
+        border <- palette[["border"]]
         text <- element_text(
             family = base_family,
             face = face,
-            colour = "black"
+            colour = fg
         )
         ## Include the grid lines.
         if (isTRUE(grid)) {
-            panel_grid_major <- element_line(colour = gray, size = 0.5)
+            panel_grid_major <- element_line(colour = border, size = 0.5)
         } else {
             panel_grid_major <- element_blank()
         }
@@ -106,8 +107,8 @@ acid_theme_light <-  # nolint
             axis_ticks <- element_blank()
             panel_border <- element_blank()
         } else {
-            axis_ticks <- element_line(colour = "black")
-            panel_border <- element_rect(colour = "black", fill = NA)
+            axis_ticks <- element_line(colour = fg)
+            panel_border <- element_rect(colour = fg, fill = NA)
         }
         theme_linedraw(
             base_size = base_size,
@@ -128,7 +129,7 @@ acid_theme_light <-  # nolint
                 panel.grid.minor = element_blank(),
                 legend.background = element_blank(),
                 legend.position = legend_position,
-                strip.background = element_rect(colour = NA, fill = "white"),
+                strip.background = element_rect(colour = NA, fill = bg),
                 strip.text = text,
                 complete = TRUE,
                 validate = TRUE
@@ -137,31 +138,37 @@ acid_theme_light <-  # nolint
 
 
 
-#' @rdname themes
+#' @describeIn themes Dark theme that has a black background and white text.\cr
+#'   Inspired by `Seurat::DarkTheme()`, with some color modifications.\cr
+#'   Useful for visualizing many points with a high dynamic color range, such
+#'   dimension reduction plots.\cr
+#'   See `darkPalette` for color values.
 #' @export
 acid_theme_dark <-  # nolint
     function() {
+        palette <- darkPalette
+        ## Set dark mode global variable that we can access inside functions.
+        options(acid.dark = TRUE)
         assert(
             isNumber(base_size),
-            ## Don't use `isString()` check on `base_family`, since empty string
-            ## is allowed by ggplot2.
-            is.character(base_family) && length(base_family) == 1L
+            is.character(base_family),
+            isScalar(base_family)
         )
         face <- match.arg(face)
         assert(isNumber(aspect_ratio, nullOK = TRUE))
         legend_position <- match.arg(legend_position)
         assert(isFlag(grid))
-        ## Set dark mode global variable that we can access inside functions.
-        options(acid.dark = TRUE)
-        gray <- "gray10"
+        bg <- palette[["background"]]
+        fg <- palette[["foreground"]]
+        border <- palette[["border"]]
         text <- element_text(
             family = base_family,
             face = face,
-            colour = "white"
+            colour = fg
         )
         ## Include the grid lines.
         if (isTRUE(grid)) {
-            panel_grid_major <- element_line(colour = gray, size = 0.5)
+            panel_grid_major <- element_line(colour = border, size = 0.5)
         } else {
             panel_grid_major <- element_blank()
         }
@@ -170,9 +177,10 @@ acid_theme_dark <-  # nolint
             axis_ticks <- element_blank()
             panel_border <- element_blank()
         } else {
-            axis_ticks <- element_line(colour = "white")
-            panel_border <- element_rect(colour = "white", fill = NA)
+            axis_ticks <- element_line(colour = fg)
+            panel_border <- element_rect(colour = fg, fill = NA)
         }
+        ## FIXME Can we consolidate with `theme_light()` better here?
         theme_linedraw(
             base_size = base_size,
             base_family = base_family
@@ -186,14 +194,14 @@ acid_theme_dark <-  # nolint
                     angle = 90L, hjust = 1L, vjust = 0.5
                 ),
                 axis.ticks = axis_ticks,
-                legend.key = element_rect(colour = NA, fill = gray),
+                legend.key = element_rect(colour = NA, fill = border),
                 legend.background = element_blank(),
                 legend.position = legend_position,
                 panel.background = element_blank(),
                 panel.border = panel_border,
                 panel.grid.major = panel_grid_major,
                 panel.grid.minor = element_blank(),
-                plot.background = element_rect(colour = NA, fill = "black"),
+                plot.background = element_rect(colour = NA, fill = bg),
                 strip.text = text,
                 complete = TRUE,
                 validate = TRUE
@@ -202,6 +210,17 @@ acid_theme_dark <-  # nolint
 
 ## Set the formals.
 formals(acid_theme_dark) <- formals(acid_theme_light)
+
+
+
+#' @describeIn themes Dark theme that uses Dracula color palette.\cr
+#'   See `draculaPalette` for color values.
+#' @export
+acid_theme_dracula <- acid_theme_dark
+## Change the first line of function body to import dracula palette.
+b <- body(acid_theme_dracula)
+b[[2L]] <- quote(palette <- draculaPalette)
+body(acid_theme_dracula) <- b